Implemented crash data reporting, though it's only some stubs on Windows and DOS.  On UNIX with USE_UNIX_SIGNALS defined, when any crash causing signal happens it will dump to a file the current crawl_state, anything caught by the items and monsters scans, and level building info if the crash happened during level generation.
Also, if crawl is linked against the GNU C library (and the exectuable is in
ELF format) it will dump the stack trace.  The code attempts to automatically
detect the presence of glibc, but that might not work on all systems.  This
should work on OS X, since there's an OS X man page for the glibc functions
that get the stack trace.  Don't know if it would work with MinGW.

Actually getting function names for the stack trace requires the use of the
"-rdynamic" linker option, which increases the size of the stripped executable
by 27% (yikes!), but still prints the function names even when stripped.

All of the function names in the stack trace are mangled C++ ones, but that
shouldn't be too much of a problem.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@8532 c06c8d41-db1a-0410-9941-cceddc491573