Add a tutorial trigger for the highly unlikely case the player encounters a monster (oklob plant, hostile statue) during the tutorial that triggers an autoexclusion.
There really needs to be an in-game way of explaining what they're for
when such an autoexclusion happens to the player for the first time,
particularly for the tiles build where excluded grids are visibly
different from non-excluded ones. I'm afraid the FAQ is too obscure for
this one, as well. Maybe add something to descriptions of monsters that
have an exclusion centered on them?


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@9831 c06c8d41-db1a-0410-9941-cceddc491573