Tiles: Save per-character equipment setting and use it in the character        selection menu.
TODO: Use this instead of dolls.txt for initialising dolls, falling back on
      dolls.txt if name.tdl cannot be found and for new characters.

I've tried displaying floor type but found it resulted in an information
overload, and also distracts from the main information (species,
equipment), esp. since it's mostly floor and most of the tile is covered
anyway. I left it in (though commented out), so it's in the repository,
but will remove it afterwards.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@10122 c06c8d41-db1a-0410-9941-cceddc491573