Remove four scroll-duplicate spells: Detect Curse, Remove Curse, Magic Mapping, and Identify. [*]
This leaves six spells in the Divinations school, of which Forescry and
See Invisible could easily be moved to other schools; Detect Secret
Doors could be removed; and Detect Traps/Items/Creatures should be moved
to some kind of misc. item with charges (whether rechargeable or not). I
didn't want to remove those until the replacement code is in.

I added a hack to make sure the probabilities of Xom "casting" magic
mapping remain the same. That will have to be repaired later,
preferrably by handling it together with Detect Items/Creatures as a
separate Xom effect.

Shifts spells in saves and bones files.

*) This is different from the recent addition of the scroll of silence
in that these four spells pose no risk whatsoever and can be spammed
safely, making even the food and magic costs negligible.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@10542 c06c8d41-db1a-0410-9941-cceddc491573