initfile.cc: improve accuracy of crawl_base variable
Windows, Mac OS X and Linux all provide handy ways to retrieve
the absolute path of the running executable. This is especially
useful since argv[0] can be a relative path.

This fixes a bug on the Mac OS X build where the app cannot
locate the graphics resources when when they're needed, so the
application bombs out instead.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>