tilesdl.cc: fix layout bug and crash with certain resolutions
This one was a pain to hunt down, but it was a surprisingly simple
fix. Can you believe an X value was used where a Y value was needed?

Anyway, the bug was that with certain resolutions (i.e. 1440x900),
the game would crash once the message list needed to be rendered.

And the layout was also completely wrong if you did something like
an abnormal resolution of 800x850, which shows the message list
near the middle of the screen instead of at the bottom.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>