files.cc, newgame.cc: support savegame paths with spaces in them
For instance, on Mac OS X, we save to this path:
   ~/Library/Application Support/Dungeon Crawl Stone Soup

But if we just pass this path to an external program via system()
without doing anything to it, the app assumes that spaces are
intended to indicate separate parameters.

We have to escape the paths in one of two ways.

UNIX (backslashes):
   ~/This\ Crappy\ Path\ Has\ Spaces\ In\ It

or Windows (doublequotes):
   "C:\This Crappy Path Has Spaces In It"

Signed-off-by: Steven Noonan <steven@uplinklabs.net>