tilesdl.cc: fix segfault caused by negative Y-value
In my case, the cause was a crawlrc with these values:
  view_max_width=80
  view_max_height=70

I did a USE_TILE build, ran it, and the value of
crawl_view.msgsz.y became negative, causing a rather nasty
segfault later in initialization.

I'm pretty sure that using 'std::max' here is not the best
solution, but it at least avoids the negative number case.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>