A rewrite of losight() to allow future generalization.
Parameters relevant to the LOS calculation are now passed as
an object los_param that takes care of coordinate translation,
bounds checking, opacity of cells and what is written to the
output array.

A check for in_bounds was changed to map_bounds for simlicity;
this should not have any effect.

Special casing for arena was removed from losight; instead
calc_show_los now calls the new losight_permissive.