Split and correct grid_see_grid.
There's now cell_see_cell which tests for visibility correctly,
albeit slowly. That's only used for placing features in the abyss
currently, but needs to be improved if it gets used more.

The second function used in monster movement was left as is, but
renamed to can_go_straight. This remains unreliable but probably
serves its purpose.

The grid_see_grid test is adapted to cell_see_cell; the
grid_see_grid lua binding is now actually useful.

Also reduce the number of tests in los_csc and los_symm.