makefiles: on Mac OS X, use the same frameworks the Xcode project uses
Before, I had it use the MacPorts libs, but this isn't the best way to do
it, even with a development build. Ideally, for debugging purposes, we
really want to be using the same libraries we'd use for a release build.

This also adds some command-line options to set the type of build:
  ARCHS: This can be set to something like "-arch i386 -arch ppc" for
         a 32-bit universal build.
  SDK_VER: The Mac OS X SDK version (i.e. 10.4, 10.5, etc)
  GCC_VER: The GCC version to compile with. For 10.4, you must use
         version 4.0, but for 10.5 and above, you may use 4.2.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>