fix 'DEBUG' macro usage consistency
Sometimes we were doing #if DEBUG and others we were doing #ifdef DEBUG.

If we mix both, we have problems: If the DEBUG macro isn't defined, the
statement '#if DEBUG' doesn't really make sense logically, because 'DEBUG'
has no value. And if we '#define DEBUG 0', then the '#ifdef DEBUG's
become true statements.

The easiest fix is to swap out the #ifs with #ifdefs.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>