New hell end maps (Neil Middleton).
Also applied the cosmetic changes (colours and descriptions for walls,
colours for floors) to the already existing branch end maps. Did not
apply those to the whole branches because it might be best to have the
branch ends stand out visually.
Someone should probably have a look at Coc:7, Geh:7, Tar:7 and Dis:7
in tiles.