Update all exclusions' LOS from travel_init_new_level().
This fixes a bug related to BR 2849608: If you created an exclusion on a
level, left that level, saved and restored off-level, and returned to
the level, the exclusion filled the complete LOS radius.

I'm not sure why it was even restricted to LOS radius, though, and
didn't fill an entire ENV_SHOW_OFFSET-size square.