makefile: Automatically optimize for your machine
Setting the makefile variable AUTO_OPT_GCC will give GCC a list of
optimization flags suited for the machine make is run on. It does this
using util/auto-opt.sh.  "-march=native -mtune=native" are always
outputted, which causes GCC to detect the CPU type you use and generate
code for that.  On Linux systems it also looks at the flags line of
/proc/cpuinfo to pick x86 optimization flags.