Add compilation information to Version namespace
The Version namespace now provides the compiler which was used, the OS,
machine type and processor type the compilation was done on, and the
CFLAGS, CFLAGS_L and LDFLAGS which were used.  If GCC was the compiler,
it also provides the compiler version.

This information is included in crash reports, and can be dumped using
the new command line option "-version".