gen_ver.pl: don't bother with MD5 checking
Comparing MD5 hashes used to be important for this script, because in the
project I developed it for, it was a globally included header (by something
akin to AppHdr.h). Preventing it from changing was pretty important. But it
doesn't matter now, since we only make it a dependency of version.cc, and
the makefile regenerates it on every build (which is handy).

Signed-off-by: Steven Noonan <steven@uplinklabs.net>