A wrapper for itemname:make_name().
Seed is always stuff:random_int(), but "all caps", "maxlen" and "start"
are specifiable, though default to false, -1, and 0 (as per
make_name()).

As each parameter is a different type (bool, int and char), it would be
good if the wrapper looked at type instead of parameter position, but
this will do for now.

Signed-off-by: Robert Vollmert <rvollmert@gmx.net>