arena.cc: Try harder to preserve respawners
If there isn't any room to place a respawning monster in, then rather
than letting it vanish, keep it on the list of pending-respawns and hope
that space clears up later.  Also, if a side with pending-respawns is
about to loose because it has 0 active members, then *make* room to
place a respawn.