Fix bug [2879187] Gods gifting duplicate books
Mark the type of a book as known when gifting it so that gods that
give books in a fixed order don't gift the same book twice if the
player never bothers to read the book the first time it is given.

I guess this is relevant because updated kiku doesn't use a gift
timeout.