New demonspawn core mechanic (from dpeg on IRC)
Demonspawn gain all mutations at level 1, and the levels increase
over time, such that at level 27 5 mutations are level 3 and the
6th is level 2.

At 17 (6 * 3 - 1) randomly chosen levellups, the PC gains one level
in a current or new mutation; the chance of raising an existing
mutation is (# of demon mutations) / 6, so the number of total
mutations will reach 6 around level 15 and stay there, but the levels
of the mutations will increase.

This is the first part of the DS rework; it is applicable as is, and
has been tested (raising 8 demonspawn to level 27 in debug mode and
sanity checking the results).  However, it is not fully balanced as
is for two reasons:

1. Not all demonspawn mutations have 3 levels yet, so you sometimes
   run out of available mutations before level 27.  When this happens,
   this patch gives a bonus mutation.

2. Some demonspawn mutations are very broken at level 3, such as fire
   resistance and teleport at will (level 3 is instantaneous).

So demonspawn will get a lot stronger until the rest of the changes
come in.

Signed-off-by: Stefan O'Rear <stefanor@cox.net>
Signed-off-by: Robert Vollmert <rvollmert@gmx.net>