Fsim improvements.
The current combat simulation system is flawed in that it only considers 2
skills for all melee attacks: fighting and the relevant weapon skill. This
has made meaningful studies of pain weapons and the enhancer staves
difficult. In addition, unarmed delay in heavy armour is very dependent on
armour skill, is not considered. The way this is implemented causes
unarmed delay to be fractional on average (you have a chance of not
lowering delay using your unarmed skill if you're in heavy armour), while
the fsim module only considers time as an integer.

This patch sets all magic skills to the same level as the weapon skill and
the armour skill equal to the fighting skill (roughly half weapon) during
the simulation. It also calculates the average time per attack as a double
instead of a long, outputting a decimal value as needed.

Signed-off-by: Robert Vollmert <rvollmert@gmx.net>