Fix bug [2864019] - assertion failure after pressing ctrl-O
An assert was getting triggered in format.cc when the line wrapping
used in display_overmap placed a line break after a text color tag
was opened and before it was closed. This was happening with the brown
text used for shop locations and also the yellow text used for branch
locations.

Using linebreak_string2 instead of linebreak_string takes care of the
shop location lines since it places line breaks on whitespace. For the
branch location strings placing the padding spaces after the </yellow>
tag instead of between the tags took care of the issue.

A more principled thing to do would be to make formatted_scroller keep
track of what tags are open across line breaks, but I got lazy.