FR 1837856: note vaults generated in dump
Include new dump section "vaults", which for game-end dumps (or any
wizard-mode dumps) will include a list (by place) of all the vaults
generated in the game.

The savefile now includes a player-associated CrawlHashTable
(you.props), which can be used for storing player information which
doesn't need to be accessed efficiently.