Electrocution discharge in water ala FR 1637214
Weapons of electrocution now discharge in water, if the target is
touching the water and not rElec, hitting all adjacent
water-touching non-rElec creatures for about half the normal
electrocution damage.

Particularly notable new code is the implementation of an
area-of-effect callback for beams, as well as a function and
structure for weapon effects that should only happen after the
target would have died, if it was going to die, and therefore
cannot safely make use of its data.

Issues that still need to be decided:
- How doe Xom feel about this? (eg. If creatures hurt themselves
  this way.)
- Should it ask you if you want to attack when you know the
  discharge will hit yourself?