Bolt of Iron->Iron Shot, Ice Bolt->Fling Icicle
Based on FR 1801710

Bolt of Iron and Ice Bolt were both mechanically different from the other
"bolts" in that they didn't penetrate targets. In addition, Ice Bolt was
very easy to confuse with Bolt of Cold. So they got their names changed to
Iron Shot and Fling Icicle, respectively. This change is both external (in
the game output) and internal (how the code refers to the spells). This
commit also changes the names of the spells in the data files (ie. for
tiles) and the names of some data files themselves (tile images).
(See the file list.)

Note that Lemuel's comment about the concept (and tile art)of "bolt" not
fitting with the way the "bolt" spells behave is still valid, but that's
another issue. I think this stems from a confusion between lightning
"bolts" and crossbow "bolts."

Signed-off-by: Eino Keskitalo <evktalo@users.sourceforge.net>