Pre- and post-savefile callbacks.
The constructor of class SavefileCallback can be used to add a callback
which is called before a game is saved and after it's restored.  This is
useful to move data in and out of you.props, or after a restore if
there's data not stored in the savefile which can be easily regenerated.

NOTE: I experienced some weirdness with C++ global constructors, and
since global constructors probably differ with compiler and system,
until this has been tested on all compiler/system combinations, any code
which uses the callbacks should fail gracefully if the callbacks aren't
called.