Replace ray_def by a version using "diamond rays".
Think of the map tiled with squares (the cells).

Previously, a ray was blocked by any part of an opaque cell
except the very corners (to allow looking through diagonal
maps -- we can move through those after all).

In the new version, the rays are the same, but a dungeon feature
is only considered to occupy the the central "diamond": convex
hull of the mid points of the edges. The corners of the
diamond are considered part of the feature, so vertically or
horizontally adjacent walls still block.

This has some effect on LOS: It's a bit more permissive on
diagonals and around corners.

The main advantage is that most rays will now have a footprint
(the sequence of cells it passes through) that is like the
old anti-aliased rays, without requiring any ray munging.

TODO:
  * Make sure we don't hang because of stopping at a corner.
  * Implement proper bouncing.
  * Reenable setting/getting degrees for chaos beams.