Documentation changes: remove +Map, expand on vault generation.
Mapping is no longer available as a randart property, so removing it
from the list. While +Tele is likewise no longer available, the unrand
Cloak of the Flash has evocable teleportation, so leaving it there for
now (randart rings of teleportation appear as *TELE).

The vault generation test set can be run individually by specifying it
on the command-line. This saves going through all the LOS tests if
generating maps with some regularity.