Remove +Tele from Cloak of Flash and documentation.
After discussion with sorear on IRC, Cloak of the Flash's evocable
teleportation will not work with recent changes, so remove it. Also
remove "mapping" and "cantelep" from the list of possible brands for
unrands, as these will no longer work.

The util/art-data.pl script will need to be updated.