Get rid of multiple-meaning "int object" in env.show.
env.show is now a class show_def that stores tagged
unions of type show_type. For the moment, there's also
env.show_los for use in LOS determination, but that
should become an array of boolean at some point.

This breaks save compatibility.

Tiles and console version build and appear to work
fine, but this kind of change is likely to have
some side-effects.