Clean up random coordinate usage a bit.
Specifically, use random_in_bounds() in a few places instead of random
ranges involving X_BOUND and Y_BOUND values.  The former won't return
any coordinates in the dungeon's map border (which acts as a wall), but
the latter will, and all the changed places are looking for empty
dungeon squares, so this should make these operations more efficient.