New unique: Aizul, disgraced guardian naga turned naga warrior.
Casts Venom Bolt, Blink Other, Ensorcelled Hibernation, Slow and heals
himself. Comes with a ruby-red naga barding, a shield, and a chance of a
very nice weapon. Can appear on Snake:4 and Snake:5, and also on similar
dungeon levels to Margery and co.

Backstory: the guardian of a horde of treasure who fell asleep on the
job. Thieves raided the place, and he was disgraced and kicked out.
Turned to the role of a warrior, seeking redemption through battle.

Currently, there's something weird with the debug build at 'xv' for
monster description not displaying the right information all the time.