New unique: Pikel, the big kobold slaver.
Appears from D:4 through D:9, and comes with a band of human slaves.
They will turn neutral (and be pacified, granting 1/2 EXP) if you
destroy Pikel before destroying them. Pikel generates with a whip
(either flaming or pain), and a black robe.

The "pikel_band_neutralise()" function is quite hacky and assumes that
all MONS_HUMANS with MF_BAND_MEMBER are part of Pikel's band. This
should be changed as soon as there is a better way of tracking who is in
what band.