Prevent chaos weapons from using SPWPN_CONFUSE on the player
SPWPN_CONFUSE is what confusing touch uses to inflict confusion. The
SPWPN_CONFUSE case in melee_attack::apply_damage_brand isn't set up to
work on the player, and when that case came up with the player
defending an assertion failure was triggered.

I made the case in apply_damage_brand do nothing if used on the player.
The case could be made more general, but it might make sense to
implement a confuse brand specifically for chaose weapons since the
HD check used for confusing touch isn't consistent with how the other
chaos-weapon-caused status effects work.