Reform SH to be more transparent
Before: SH was a function of items and skills.  Blocking was a function
of SH, skills, and stats.

Now: SH is a function of items, skills, and stats.  Blocking is a
function of SH and SH alone.

Your ability to block attacks in given equipment should be more or less
the same, although there will be small differences.  Condensation Shield
was doubled (since nobody uses it), Divine Shield was made to not use
intelligence.

SH values have been rescaled to be on approximately the same scale as
AC and EV.  This meant increasing the max enchant of shields to +3.