Tweakes to uniques.des: Lair, add Gastronok, fix Grum.
Lair:8 can be relatively as deep as 20, and as low as 8, so mark any
unique spawning between these levels as !Lair. Unfortunately,
default-depth doesn't stack with DEPTH, so it is not possible to use
this to automatically exclude Lair as a possible branch.

Also fix Grum's entry, which somehow got merged with Gastronok's, the
latters being deleted.

Finally, remove old code from the end of the file, as it is unneeded.