Add an is_holy() function to the actor interface.
This is consistent with similar functions in that it checks not only for
MH_HOLY holiness, but whether the monster is a priest of a good god
(currently, there are none) and whether it uses holy spells (currently,
there are none after the changes described below).

Minor and Major Healing are now enchantments rather than holy spells.
The beam code treats them as such; wands of healing aren't holy items
the way e.g. wands of draining are evil items; and unholy monsters
(among others, Lom Lobon!) can cast these spells.