Remove ogre from shrine_trog_oka.
Berserk ogre from a D:2-accessible Shrine is instant death for most
characters; orc warriors are not much better, so I've added in a normal
orc (which is still quite scary while berserk) as well. Also tweaked the
weapon placement slightly, and moved the exit to be closer to the entry,
allowing for quicker escape if you don't want to face the threat.