Tweak Dowan/Duvessa death effects + invisibility (jpeg).
Should now work even if one (the one being killed) or the other (the
living twin) or bother are invisible. Add a new subset of flavour
message to cope with this ("It weeps, then rights herself and shakes her
weapon" doesn't work with invisible).

Still needs to be tweaked, as Dowan will flee/Duvessa will go berserk if
one of them is stabbed. Also, flavour messages may need to be adjusted
to take into consideration trap deaths.