Place some of the altars outside of the Temple
Framework for placing altars outside the Temple, into what I call
"overflow temples".  It currently places twelve altars in the Ecumenical
Temple and two outside at D:2 through D:9, with each of the two overflow
altars being placed in a "temple" which consists of a bare altar.  The
overflow temples can actually contain any number of altars, but I'm not
a vault designer, so I went the minimalistic route.

The levels the overflow temples are placed on can be listed with the
wizard command "&:".