dat/entry.des: elemental_entry excludes/warnings
Remove vault-placed exlcudes from the elemental entry vault, and instead
put warnings on the secret doors.  When the secret doors pop open after
they're detected, the player will be notified of the opening-prompt
placed on them, and asked if they wish to put a travel-exclude on top of
the door.  If users ignore that warning and wander in anyways, they were
adequetly warned.