Change ballistomycete spore mechanics again
New status: if a ballistomycete or giant spore dies any ballistos on
the level get +1 to a counter. Ballistos with a count higher than zero
get spore production on a short (~150 turn) timer. When a spore is
spawned (by a ballisto) the count of all ballistos on the level is
decreased by 1.