religion.cc: mons_make_god_gift(): switching gods
mons_make_god_gift() would assert if a monster already had mon->god set
to something besides GOD_NO_GOD or the god doing the gifting, but this
didn't work with Jiyva slimifying orcs that worship Beogh.  I think this
is left over from before there was MF_GOD_GIFT, when
mon->god != GOD_NO_GOD mean that the monster was a god-gift.  So now
instead it asserts if the monster is already a god gift.