Leave viewwindow early when not drawing.
viewwindow has code to not draw if the player is travelling (depending
on travel_delay etc). This check is now moved before the code that
updates buffy for non-tiles builds. I'm not sure whether the tiles
drawing code doesn't have important side-effects -- if not, the
tiles check could be moved before the main loop as well.