Change ballistomycete spawn mechanics (again)
When a spore pops or a ballistomycete is killed randomly select a
single ballisto to activate instead of activating the entire level's
worth. Along with this, only deactivate the ballisto that produced a
spore instead of deactivating the entire level.

Rationale: since inactive ballistos don't interrupt travel anymore it's
nicer from an interface perspective to keep most of the ballistos on
the level inactive.