Revert removal of the speed brand, pending discussion.
Such grand changes should not be made out of the blue.  There was not a single
mention of it before, save a line in a long brainstorm FR.  It has a large
impact (badly hurting short blade users, making axes far more a non-brainer
over long blades/polearms, and so on).  We may decide this is the way to go,
but a discussion needs to be done first.

And, re-reverting it is a single command so no work is lost...