Fix an issue with noise generation in melee combat
Noise generated by attacking in melee range is handled by
melee_attack::handle_noise. handle_noise was using defender->pos() as
the point of origin for the noise, but defender->pos() is not reliable
in some situations, if a monster is killed its position is reset to
zero. I added a coord_def parameter to handle_noise to use as the point
of origin instead of trying to use defender->pos and made all calls to
handle_monster store the defender's position prior to damage being
applied.

Aside from creating noise in the wrong area calling noisy with a
position of (0, 0) can cause an assertion failure at mon-behv.cc,
line 1609 in some situations.