handle_behaviour(): guess pos of invis foes
If a monster didn't know the position of an unseen but nearby foe, then
casting a spell against the foe would cause an assert.  So now the
monster tries to guess the position of the foe: a random square within
radius 3 of the foe, or radius 2 if the monster can sense invisible.
This algorithm no doubt needs improvement.