Travel exclude speedup
Speed up is_excluded() by making the class exclude_set, which caches the
set of excluded points, calculating the set only when exclusions are
added or removed (or when an old level is loaded).

NOTE: This recomputes the set of excluded points once upon entering
level-map mode and once upon exiting it.  If this takes to long on slow
machines, this can be improved.