Disable tutorial waking monsters.
Tutorial shouldn't have side-effects; also, the "may stab" brand
can still be explained later.

In console debug + tutorial, this was causing an assertion. Briefly:

Handle_seen_interrupt issues a monster "newly seen" interrupt, which
calls tutorial_first_monster, which in turn alerts the monster, which
causes it to shout, which calls learned_something_new(MONSTER_SHOUT),
which prints a message, which issues a message interrupt, which
issues another message, which calls flush_comes_into_view, which
interrupts with the same monster again, but now "just seen", which
causes an assertion.