Avoid triggering an assert when firing while standing on a plant
An assert in ray_blocker was being triggered off of a call to
_get_monster_desc when attempting to cast targeted spells while
standing on a plant (as a Fedhas worshiper).
find_ray returns false when the target square is equal to the source,
which triggers the assert in ray_blocker. I added a check for the
monster being on the same square as the player prior to the call to
_blocked_ray in _get_monster_desc, although this may not be the right
way to deal with the issue.