Un-underpopulate '4's by adding sixfirhy: elec theme, unorthodox timings.
Sixfirhies move in bursts: speed 30, but act only on 4 turns out of 12.
They also get a bonus to moving compared to other actions (move cost 60%).

Damage brand is AF_ELEC -- it's unproportionately dangerous, so the numbers
only look small.

Also, electricity heals them!  It won't bring the back from over the edge
if physical part of damage got them to 0 hp or less, in this case, they'll
explode instead.

Keeping with the long tradition of Crawl's demon names, the credit for this
one goes to timecircuits' cat.