Remove Zin's summoning-based wrath, as it no longer fits.
First, Jiyva considers eyeballs slimes for his purposes.  Second, insect
plagues are more suited to old Zin, who coupled them with angels for an
Old Testament theme, so to speak.  (The latter is most closely suited to
Fedhas, or it would be if he were the god of nature in general instead
of just plants and fungi.  It's also inappropriate for Zin now, since
one of the insects summoned causes disease and is, hence, chaotic.)