Prompt before auto-exploring the Labyrinth.
The wording may require work. I'm not sure if this is acceptable, but
the reasoning is that hitting 'o' by accident in the Labyrinth,
especially when you have travel_delay = -1, can take you from just near
the exit to miles away, and you have to start all over again.

There are instances where auto-explore may be useful in the Labyrinth,
so don't outright disable it.