Alt-L-Click to zap wand
Alt-L-Click on either the player or a monster will offer a menu of wands
to zap, and then zap it.  Only appropriate identified wands will be
offered in the menu: healing/haste/etc for the player and non-hostile
monsters, and attack wands for hostile monsters (unidentified wands will
be ofered for all targets).  Player targetting still doesn't work quite
right yet.

Evoking rods will be added in the future.