Move Tomb branch flag no_tele_control from crypt.des to branch.cc.
This fixes shaft traps being generated on Tomb:1; apparently,
BFLAGS: was being interpreted after placing traps on the level;
shaft traps are not placed in no-telecontrol levels.

A better fix would be to make sure BFLAGS is called earlier.
Alternatively, BFLAGS could be remove -- it doesn't appear
to be used any more.