Fix bugs 2900276 and 2901145, monster beams hitting their thrower
Switch to the YOU_KILL and MON_KILL macros for determining whether a
beam was fired by a monster or the player (used to decide whether a
beam can hit a monster or the player at its point of origin). This
fixes manticores not using their ability, and also stops electric eels
(and electric eel pallet swaps) from shooting themselves.