A nasty hack to allow access to the item library in vaults.
As the item library is aptly named "item", and the function for
specifying item slots in a vault is likewise named "item", there's no
way for vaults to access any of the functionality in the item library
(it just returns errors about accessing a function value).

I'm not sure what changing the name of the item library will do,
however, so I've left it the same and simply added a second copy using
the library name of "items". It seems to function correctly, and
hopefully will not cause any issues. Perhaps the current "item" library
could actually be renamed, but this may have impact on user-defined
scripts.