Rebalance distracted stabs (FR 2902035).
For any of the stab attempts with a roll, the chance to stab
is increased by a factor 3/2 (now around dex+stab in 100
instead of dex+stab in 150), and damge bonus is halved.

This may still improve damage output of highly-developed
stabbers if they still do enough to kill most monsters
in one hit, but I think it's ok.

For the record, such a stabber will know do around 80-90
damage per hit against a confused monster, less against
a distracted one.