Fix wiz-mode Ziggurats not being properly initialised [BR2903984]
The code references the "portal" object (created in ziggurat_portal,
which is also passed to the onclimb initialiser) which is not actually
used for Wizard Mode portals.

Fixed this by a) checking whether or not there was a portal included in
the zig() "check to see if we're using &P", and also adding a parameter
to to ziggurat_portal which just returns the actual portal.