Don't link tiles against ncurses nor libpng.
The former is a "duh" case, the latter because Crawl itself doesn't depend
on libpng, it's needed by libsdl-image only.  Such unneeded dependencies
can then result in waste (if all libraries use proper symbol versioning)
or crashes (if not) once sdl is rebuilt against a different version of libpng.
(Suggestions I'm doing this not to avoid future crashes but merely because
Debian tools are screaming at me are evil lies, honest! :þ).