Simplify scrolling handling in message_out.
The newline parameter to message_out is gone. Instead, a
pointer to the message line is passed which may lie outside
the message window. The window is then scrolled to make
that line the last line.

This also updates libnonunix.cc, but I haven't been able
to test those changes.