Add an option to draw the player cursor manually.
Set use_fake_player_cursor = true to use this. This allows
defaulting the cursor to off, which means less cursor blinking
say when travelling.

This patch also moves the cursor control closer to where it's
used, so that the cursor is only turned on for actual input
prompts. I've definitely missed some prompts here: almost
anything that prompts for a single character and reads it
using getch() will not have a cursor with use_fake_player_cursor
set at the moment.