Hacked uniq_crazy_yiuf_cottage a bit.
It was regularly splitting levels up. Despite the fact that it no longer
clobbers stairs, it's still placed after the connectivity confirmation,
so it can still split the level. I've added paths through it and entry
points, so hopefully this should solve the issue.

May need to be reverted if it doesn't fix the issue, or causes other
problems.