New Lua iterators: point_iterator, slave_iterator, tweaks.
The first (point_iterator) accepts a table of points (such as the ones
return by dgn.find_marker_positions_by_prop, or any other function which
returns points) and is filterable and stateful.

The second is a quick wrapper over dgn.find_marker_positions_by_prop.

Tweaked rectangle_iterator to supply an "rvi" (return value instead)
which will store the value of the filter instead of expecting the filter
to return a true or false value. Setting this to false will behave as
before; setting it to true allows us to have the filter alter the point
(mons_at, items_at, marker_at_pos, etc).