Optionally display message window at top of terminal.
Set messages_at_top=true to use.

This is not quite aesthetically pleasing since character name
and title which form a kind of heading for the screen are now
in the middle. It might be better to display them in an extra
line across the top.

Also, the layout should really be moved out to lua and made
completely user configurable.