New monster spec tags for marking monsters summoned.
dur:1-6, marks the abjuration duration of a monster. sum:<str>, any of
"clone", "animate", "chaos", "miscast", "zot", "wrath" or "aid", or the
relevant spell we're mimicking; nas:<str> equivalent to the non-actor
summoner.

Not setting the non-actor summoner will mean that kills by these pseudo-
summons are unmarked as such.