Unique placement-related function find_maps_for_tag.
Generates a vector of map_defs that have the relevant tag. Can check
depth (defaults to false) as well as whether or not a vault has
previously been placed (defaults to true).

Convenience function weight_map_vector returns an integer corresponding
to the total weight of all of the maps contained within the vector.
These functions will be used to re-do unique placement as per dpeg's
design (see c-r-d email).