Massively expand tile functionality in vault definitions.
This commit creates a new specifier for vaults: "TILE". Used much in the
same way as COLOUR, it can apply any specific tile to a feature. Example
syntax is specified in the syntax file, but copied here for clarity:

TILE: x = wall_flesh
      Identical to FTILE and RTILE in syntax, but closer to COLOUR in
      functionality. Instead of replacing the floor or relevant rock
      tiles, this can be used to replace the tile used for any specific
      feature.

      This can be used in combination with FTILE and RTILE to change
      the appearance of features. It can only be used with previously
      specified tiles, however.

      Like COLOUR and FTILE, this should be used sparingly and to good
      effect.

Please, feel free to update vaults to use this! We want to ensure that
tiles players get the same experience as ASCII players do. This is only
the first stage in a push for greater flexibiltiy through tiles, but
hopefully it'll have a good impact.