Add more tiles for Cigotuvi WizLab.
This includes deformed human (sickly green, with a single wing), a deformed
elf (horrible fleshy grey colour with a hole in the middle), and a
deformed orc (missing an arm and part of its face).

Also, add a tile for "flesh golem". This is basically a reworked stone
golem, using some colours from the wall_flesh series of tiles. Might be
a bit dark, but it looks okay at the minute.